requirements.txt
# FastAPI and Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain and AI
langchain>=0.1.0
langchain-community>=0.0.13
langchain-google-genai>=0.0.6
langchain-openai>=0.0.5
google-generativeai>=0.3.2

# Vector Store
faiss-cpu>=1.7.4

# Dataset
datasets>=2.16.0
huggingface-hub>=0.20.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# Utilities
httpx>=0.26.0
tenacity>=8.2.3
PyMuPDF>=1.23.0
